refactored so that each url req points to a specific controller file relevant to its request
